pyproject.toml
[tool.poetry]
name = "dfa_test"
version = "0.1.0"
description = ""
authors = ["Герман Гриценко <gerfest.gricenko@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
python-decouple = "^3.6"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
drf-yasg = "^1.21.3"
djangorestframework-simplejwt = "^5.2.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[flake8]
ignore = "F811"
max-line-length = 101

[tool.isort]
profile = "black"
skip = [".gitignore", ".dockerignore", "__init__.py"]
